+++ src/HaystackInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Haystack;
+
+use Haystack\Container\ContainerInterface;
+use Haystack\Functional\FunctionalInterface;
+use Haystack\Math\MathInterface;
+
+interface HaystackInterface extends \IteratorAggregate, \ArrayAccess, \Serializable, \Countable, ContainerInterface, FunctionalInterface, MathInterface
+{
+
+}